requirements.txt
python-dotenv
pymongo
fastapi
pydantic
pydantic[email]
uvicorn